Search API perfomance DB optimization - Default Dispatcher
