+++ engine/src/main/java/com/google/android/fhir/db/impl/DatabaseImpl.kt
@@ -21,6 +21,7 @@ import androidx.annotation.VisibleForTesting
 import androidx.room.Room
 import androidx.room.withTransaction
 import androidx.sqlite.db.SimpleSQLiteQuery
+import ca.uhn.fhir.context.FhirContext
 import ca.uhn.fhir.parser.IParser
 import ca.uhn.fhir.util.FhirTerser
 import com.google.android.fhir.DatabaseErrorStrategy
@@ -37,12 +38,15 @@ import com.google.android.fhir.db.impl.entities.LocalChangeEntity
 import com.google.android.fhir.db.impl.entities.ResourceEntity
 import com.google.android.fhir.index.ResourceIndexer
 import com.google.android.fhir.logicalId
-import com.google.android.fhir.pmap
 import com.google.android.fhir.search.SearchQuery
 import com.google.android.fhir.toLocalChange
 import com.google.android.fhir.updateMeta
 import java.time.Instant
 import java.util.UUID
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.async
+import kotlinx.coroutines.awaitAll
+import kotlinx.coroutines.coroutineScope
 import org.hl7.fhir.r4.model.IdType
 import org.hl7.fhir.r4.model.Resource
 import org.hl7.fhir.r4.model.ResourceType
@@ -229,7 +233,10 @@ internal class DatabaseImpl(
   ): List<ResourceWithUUID<R>> {
     return db.withTransaction {
       resourceDao.getResources(SimpleSQLiteQuery(query.query, query.args.toTypedArray())).pmap {
-        ResourceWithUUID(it.uuid, iParser.parseResource(it.serializedResource) as R)
+        ResourceWithUUID(
+          it.uuid,
+          FhirContext.forR4Cached().newJsonParser().parseResource(it.serializedResource) as R,
+        )
       }
     }
   }
@@ -244,7 +251,8 @@ internal class DatabaseImpl(
           ForwardIncludeSearchResult(
             it.matchingIndex,
             it.baseResourceUUID,
-            iParser.parseResource(it.serializedResource) as Resource,
+            FhirContext.forR4Cached().newJsonParser().parseResource(it.serializedResource)
+              as Resource,
           )
         }
     }
@@ -260,7 +268,8 @@ internal class DatabaseImpl(
           ReverseIncludeSearchResult(
             it.matchingIndex,
             it.baseResourceTypeAndId,
-            iParser.parseResource(it.serializedResource) as Resource,
+            FhirContext.forR4Cached().newJsonParser().parseResource(it.serializedResource)
+              as Resource,
           )
         }
     }
@@ -459,6 +468,11 @@ internal class DatabaseImpl(
 
     @VisibleForTesting const val DATABASE_PASSPHRASE_NAME = "fhirEngineDbPassphrase"
   }
+}
+
+/** Implementation of a parallelized map */
+suspend fun <A, B> Iterable<A>.pmap(f: suspend (A) -> B): List<B> = coroutineScope {
+  map { async(Dispatchers.Default) { f(it) } }.awaitAll()
 }
 
 internal data class DatabaseConfig(
